Convert document body elements too.
When we find a custom-style include in a document, in order to properly account for the order of application of @apply css mixins we need to insert body elements in deferred script tags as well.

This is super ugly, but it's the best amelioration of https://github.com/Polymer/polymer-modulizer/issues/154 that I've found so far.