Decouple updating test source repos from regenerating expected results.
Previously, the "build:fixtures" NPM script would both pull the latest
master commit for each repo, and regenerate the expected result.

Now, these two steps are separated into "update-fixture-sources" and
"update-fixture-expected" scripts. This makes it much easier to separate
the effect of modulizer changes vs changes to the source repo, and is
also faster to run during development.

Additionally, regenerating the expected result is now done by running
the integration test itself in a special mode via the UPDATE_EXPECTED
environment variable. Previously it was done by a separate call to
convertPackage, which meant it was easy for the configuration we use for
running the tests to get out of sync with the configuration we use for
regenerating the expected results (this seems to have actually happened
recently).